mcp-server/requirements.txt
mcp>=1.0.0
python-pptx>=1.0.0
google-genai>=0.8.0
